Refresh the QA database using a rake task
Drop and recreate the database. Load courses from Find and open them as
we do in setup_local_dev_data.

Call out to an external URL to get a list of DSI UIDs. Create support users
for those UIDs. When these users log in, the emails will be corrected to
those in DfE Sign-in.