ProviderPermissionsListComponent works with ProviderPermissions
When we get the data we need for provider_interface/provider_users#show,
use two new scopes on the ProviderPermissions model to get us the
permissions we care about, and work outwards from those.

There's no need for an intermediate representation of the available
permissions - we can ask the model.